refactor tests/appsTest file
