+++ tests/pactHelper.js
@@ -1,6 +1,4 @@
 var config = require('./config/config.json')
-const utf8 = require('utf8')
-
 var validUserPasswordHash = btoa(config.username + ':' + config.password)
 const Pact = require('@pact-foundation/pact-web')
 
@@ -348,184 +346,6 @@ function setGeneralInteractions (provider) {
       }
     }
   }))
-  promises.push(provider.addInteraction({
-    uponReceiving: 'a GET request for a list of apps',
-    withRequest: {
-      method: 'GET',
-      path: Pact.Matchers.term({
-        matcher: '.*\\/ocs\\/v1\\.php\\/cloud\\/apps$',
-        generate: '/ocs/v1.php/cloud/apps'
-      }),
-      headers: validAuthHeaders
-    },
-    willRespondWith: {
-      status: 200,
-      headers: xmlResponseHeaders,
-      body: '<?xml version="1.0"?>\n' +
-        '<ocs>\n' +
-        ocsMeta('ok', '100') +
-        ' <data>\n' +
-        '  <apps>\n' +
-        '   <element>workflow</element>\n' +
-        '   <element>files</element>\n' +
-        '  </apps>\n' +
-        ' </data>\n' +
-        '</ocs>\n'
-    }
-  }))
-
-  // a request to GET attributes of an app
-  const attributes = {
-    attr1: {
-      attrExists: true,
-      value: 'value1'
-    },
-    'attr+plus space': {
-      attrExists: true,
-      value: 'value+plus space and/slash'
-    },
-    属性1: {
-      attrExists: true,
-      value: '值对1'
-    },
-    'attr ': {
-      attrExists: false,
-      value: ''
-    },
-    'attr+plus space ': {
-      attrExists: false,
-      value: ''
-    },
-    '属性1 ': {
-      attrExists: false,
-      value: ''
-    },
-    '': {
-      attrExists: false,
-      value: {
-        attr1: 'value1',
-        'attr+plus space': 'value+plus space and/slash',
-        属性1: '值对1'
-      }
-    },
-    'attr1-no-value': {
-      attrExists: true,
-      value: ''
-    },
-    'attr+plus space-no-value': {
-      attrExists: true,
-      value: ''
-    },
-    '属性1-no-value': {
-      attrExists: true,
-      value: ''
-    }
-  }
-  for (const attribute in attributes) {
-    // default no data
-    let data = ' <data/>\n'
-
-    // no attributes specified, return all attributes
-    if (attribute === '') {
-      data = ' <data>\n'
-      for (const [key, value] of Object.entries(attributes[attribute].value)) {
-        data = data +
-          '  <element>\n' +
-          '   <key>' + utf8.encode(key) + '</key>\n' +
-          '   <app>someAppName</app>\n' +
-          '   <value>' + utf8.encode(value) + '</value>\n' +
-          '  </element>\n'
-      }
-      data = data + ' </data>'
-    } else if (attributes[attribute].attrExists === true) {
-      // attribute exists
-      data = ' <data>\n' +
-        '  <element>\n' +
-        '   <key>' + utf8.encode(attribute) + '</key>\n' +
-        '   <app>someAppName</app>\n' +
-        '   <value>' + utf8.encode(attributes[attribute].value) + '</value>\n' +
-        '  </element>\n' +
-        ' </data>\n'
-    }
-    promises.push(provider.addInteraction({
-      uponReceiving: 'a request to GET ' + attribute + '-attribute of an app',
-      withRequest: {
-        method: 'GET',
-        path: Pact.Matchers.term({
-          matcher: '.*\\/ocs\\/v1\\.php\\/privatedata\\/getattribute\\/someAppName\\/?' + encodeURIComponent(utf8.encode(attribute)) + '$',
-          generate: '/ocs/v1.php/privatedata/getattribute/someAppName/' + encodeURIComponent(utf8.encode(attribute))
-        }),
-        headers: validAuthHeaders
-      },
-      willRespondWith: {
-        status: 200,
-        headers: xmlResponseHeaders,
-        body: '<?xml version="1.0"?>\n' +
-          '<ocs>\n' +
-          ocsMeta('ok', '100') +
-          data +
-          '</ocs>'
-      }
-    }))
-  }
-
-  // a request to GET attributes of an app that has no values set
-  let data = ' <data>\n'
-  const values = ['attr1', 'attr+plus space', '属性1']
-  for (let i = 0; i < values.length; i++) {
-    data = data +
-      '  <element>\n' +
-      '   <key>' + utf8.encode(values[i]) + '</key>\n' +
-      '   <app>someAppName-no-value</app>\n' +
-      '   <value></value>\n' +
-      '  </element>\n'
-  }
-  data = data + ' </data>'
-  promises.push(provider.addInteraction({
-    uponReceiving: 'a request to GET attributes of an app that has no values set',
-    withRequest: {
-      method: 'GET',
-      path: Pact.Matchers.term({
-        matcher: '.*\\/ocs\\/v1\\.php\\/privatedata\\/getattribute\\/someAppName-no-value$',
-        generate: '/ocs/v1.php/privatedata/getattribute/someAppName-no-value'
-      }),
-      headers: validAuthHeaders
-    },
-    willRespondWith: {
-      status: 200,
-      headers: xmlResponseHeaders,
-      body: '<?xml version="1.0"?>\n' +
-        '<ocs>\n' +
-        ocsMeta('ok', '100') +
-        data +
-        '</ocs>'
-    }
-  }))
-
-  const requests = ['POST', 'DELETE']
-  for (let i = 0; i < requests.length; i++) {
-    const action = (requests[i] === 'POST') ? 'enable' : 'disable'
-    promises.push(provider.addInteraction({
-      uponReceiving: action + ' apps',
-      withRequest: {
-        method: requests[i],
-        path: Pact.Matchers.term({
-          matcher: '.*\\/ocs\\/v1\\.php\\/cloud\\/apps\\/.+$',
-          generate: '/ocs/v1.php/cloud/apps/files'
-        }),
-        headers: validAuthHeaders
-      },
-      willRespondWith: {
-        status: 200,
-        headers: xmlResponseHeaders,
-        body: '<?xml version="1.0"?>\n' +
-          '<ocs>\n' +
-          ocsMeta('ok', '100') +
-          ' <data/>\n' +
-          '</ocs>'
-      }
-    }))
-  }
 
   promises.push(provider.addInteraction({
     uponReceiving: 'a request to delete a user',
